Check if folder exists before replacing html

+++ apache/docker-entrypoint.sh
@@ -4,8 +4,10 @@
 rm -f /var/www/git-wordpress/.git/index.lock
 
 # Link default HTML folder to WordPress
-rm -df /var/www/html
-ln -s /var/www/git-wordpress/html /var/www/html
+if [ -d "/var/www/git-wordpress/html" ]; then
+  rm -df /var/www/html
+  ln -s /var/www/git-wordpress/html /var/www/html
+fi
 
 # Run a script, without waiting for it to finish
 # https://unix.stackexchange.com/questions/86247/what-does-ampersand-mean-at-the-end-of-a-shell-script-line#answer-86253
